requirements.txt
accelerate==0.26.1
anyio==4.2.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==41.0.7
diffusers==0.25.1
dill==0.3.7
exceptiongroup==1.2.0
filelock==3.9.0
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.2
idna==3.4
importlib-metadata==7.0.1
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.15
networkx==3.0
numpy==1.24.1
packaging==23.2
pathos==0.3.1
Pillow==9.3.0
pox==0.3.3
ppft==1.7.6.7
psutil==5.9.8
pycparser==2.21
PyExecJS==1.5.1
python-telegram-bot==13.15
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.1
tqdm==4.66.1
transformers==4.36.2
translators==5.8.9
typing_extensions==4.4.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.13
zipp==3.17.0
